refactor(ui): add screen on ui package
