+++ app/src/androidTest/java/flyve/org/admindashboard/SplashTest.java
@@ -32,7 +32,7 @@ import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.view.WindowManager;
 
-import org.flyve.admin.dashboard.SplashActivity;
+import org.flyve.admin.dashboard.ui.SplashActivity;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
