+++ app/src/main/AndroidManifest.xml
@@ -20,20 +20,20 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".DashboardAgent"
+        android:name=".ui.DashboardAgent"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="org.flyve.admin.dashboard.SplashActivity">
+        <activity android:name="org.flyve.admin.dashboard.ui.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="org.flyve.admin.dashboard.MainActivity">
+        <activity android:name="org.flyve.admin.dashboard.ui.MainActivity">
         </activity>
     </application>
