+++ app/src/main/java/org/flyve/admin/dashboard/ui/DashboardFragment.java
@@ -1,10 +1,12 @@
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import org.flyve.admin.dashboard.R;
 
 public class DashboardFragment extends Fragment {
