+++ app/src/main/java/org/flyve/admin/dashboard/ui/DeviceFragment.java
@@ -1,4 +1,4 @@
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -14,46 +14,46 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ProgressBar;
 
-import org.flyve.admin.dashboard.adapter.FileAdapter;
-import org.flyve.admin.dashboard.adapter.FileTouchHelper;
+import org.flyve.admin.dashboard.R;
+import org.flyve.admin.dashboard.adapter.DeviceAdapter;
+import org.flyve.admin.dashboard.adapter.DeviceTouchHelper;
 import org.flyve.admin.dashboard.utils.FlyveLog;
 import org.json.JSONArray;
 import org.json.JSONObject;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class FileFragment extends Fragment {
+public class DeviceFragment extends Fragment {
 
     private ProgressBar pb;
     private RecyclerView lst;
     private List<HashMap<String, String>> data;
-    private FileAdapter mAdapter;
+    private DeviceAdapter mAdapter;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_file, container, false);
+        View v = inflater.inflate(R.layout.fragment_device, container, false);
 
         pb = v.findViewById(R.id.pb);
         pb.setVisibility(View.VISIBLE);
 
         lst = v.findViewById(R.id.lst);
 
-        LinearLayoutManager llm = new LinearLayoutManager(FileFragment.this.getActivity());
+        LinearLayoutManager llm = new LinearLayoutManager(DeviceFragment.this.getActivity());
         lst.setLayoutManager(llm);
 
         lst.setItemAnimator(new DefaultItemAnimator());
-        lst.addItemDecoration(new DividerItemDecoration(FileFragment.this.getContext(), DividerItemDecoration.VERTICAL));
+        lst.addItemDecoration(new DividerItemDecoration(DeviceFragment.this.getContext(), DividerItemDecoration.VERTICAL));
 
         // adding item touch helper
         // only ItemTouchHelper.LEFT added to detect Right to Left swipe
         // if you want both Right -> Left and Left -> Right
         // add pass ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT as param
-        ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new FileTouchHelper(0, ItemTouchHelper.LEFT, new FileTouchHelper.RecyclerItemTouchHelperListener() {
+        ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new DeviceTouchHelper(0, ItemTouchHelper.LEFT, new DeviceTouchHelper.RecyclerItemTouchHelperListener() {
 
             /**
              * callback when recycler view is swiped
@@ -62,9 +62,9 @@ public class FileFragment extends Fragment {
              */
             @Override
             public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction, int position) {
-                if (viewHolder instanceof FileAdapter.DataViewHolder) {
+                if (viewHolder instanceof DeviceAdapter.DataViewHolder) {
                     // get the removed item name to display it in snack bar
-                    String name = data.get(viewHolder.getAdapterPosition()).get("name");
+                    String name = data.get(viewHolder.getAdapterPosition()).get("UserRealName");
 
                     // backup of removed item for undo purpose
                     final HashMap<String, String> deletedItem = data.get(viewHolder.getAdapterPosition());
@@ -75,7 +75,7 @@ public class FileFragment extends Fragment {
 
                     // showing snack bar with Undo option
                     Snackbar snackbar = Snackbar
-                            .make(FileFragment.this.getView(), " removed from list", Snackbar.LENGTH_LONG);
+                            .make(DeviceFragment.this.getView(), name + " removed from list", Snackbar.LENGTH_LONG);
                     snackbar.setAction("UNDO", new View.OnClickListener() {
                         @Override
                         public void onClick(View view) {
@@ -91,6 +91,7 @@ public class FileFragment extends Fragment {
         });
         new ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(lst);
 
+
         load(loadJSONFromAsset());
 
         return v;
@@ -99,7 +100,7 @@ public class FileFragment extends Fragment {
     public String loadJSONFromAsset() {
         String json = null;
         try {
-            InputStream is = getActivity().getAssets().open("json/files.json");
+            InputStream is = getActivity().getAssets().open("json/agents.json");
             int size = is.available();
             byte[] buffer = new byte[size];
             is.read(buffer);
@@ -126,17 +127,24 @@ public class FileFragment extends Fragment {
                 HashMap<String, String> c = new HashMap<>();
 
                 c.put("type", "data"); // clasify the item if data or header
-                c.put("name", obj.getString("PluginFlyvemdmFile.name"));
-
-                long fileKb = Long.parseLong(obj.getString("PluginFlyvemdmFile.filesize")) / 1024;
-                c.put("size", String.valueOf(fileKb) + " KB");
+                c.put("id", obj.getString("PluginFlyvemdmAgent.id"));
+                c.put("email", obj.getString("PluginFlyvemdmAgent.name"));
+                c.put("fleetName", obj.getString("PluginFlyvemdmAgent.PluginFlyvemdmFleet.name"));
+                c.put("ComputerId", obj.getString("PluginFlyvemdmAgent.Computer.id"));
+                c.put("ComputerSerial", obj.getString("PluginFlyvemdmAgent.Computer.serial"));
+                c.put("UserId", obj.getString("PluginFlyvemdmAgent.Computer.User.id"));
+                c.put("FleetId", obj.getString("PluginFlyvemdmAgent.PluginFlyvemdmFleet.id"));
+                c.put("lastContact", obj.getString("PluginFlyvemdmAgent.last_contact"));
+                c.put("UserRealName", obj.getString("PluginFlyvemdmAgent.Computer.User.realname"));
+                c.put("AgentVersion", obj.getString("PluginFlyvemdmAgent.version"));
+                c.put("isOnline", obj.getString("PluginFlyvemdmAgent.is_online"));
 
                 data.add(c);
             }
 
             pb.setVisibility(View.GONE);
 
-            mAdapter = new FileAdapter(data);
+            mAdapter = new DeviceAdapter(data);
             lst.setAdapter(mAdapter);
 
         } catch (Exception ex) {
