+++ app/src/main/java/org/flyve/admin/dashboard/ui/FleetFragment.java
@@ -1,4 +1,4 @@
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -14,8 +14,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ProgressBar;
 
-import org.flyve.admin.dashboard.adapter.UserAdapter;
-import org.flyve.admin.dashboard.adapter.UserTouchHelper;
+import org.flyve.admin.dashboard.R;
+import org.flyve.admin.dashboard.adapter.FleetAdapter;
+import org.flyve.admin.dashboard.adapter.FleetTouchHelper;
 import org.flyve.admin.dashboard.utils.FlyveLog;
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -26,34 +27,34 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class UserFragment extends Fragment {
+public class FleetFragment extends Fragment {
 
     private ProgressBar pb;
     private RecyclerView lst;
     private List<HashMap<String, String>> data;
-    UserAdapter mAdapter;
+    private FleetAdapter mAdapter;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_user, container, false);
+        View v = inflater.inflate(R.layout.fragment_fleet, container, false);
 
         pb = v.findViewById(R.id.pb);
         pb.setVisibility(View.VISIBLE);
 
         lst = v.findViewById(R.id.lst);
 
-        LinearLayoutManager llm = new LinearLayoutManager(UserFragment.this.getActivity());
+        LinearLayoutManager llm = new LinearLayoutManager(FleetFragment.this.getActivity());
         lst.setLayoutManager(llm);
 
         lst.setItemAnimator(new DefaultItemAnimator());
-        lst.addItemDecoration(new DividerItemDecoration(UserFragment.this.getContext(), DividerItemDecoration.VERTICAL));
+        lst.addItemDecoration(new DividerItemDecoration(FleetFragment.this.getContext(), DividerItemDecoration.VERTICAL));
 
         // adding item touch helper
         // only ItemTouchHelper.LEFT added to detect Right to Left swipe
         // if you want both Right -> Left and Left -> Right
         // add pass ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT as param
-        ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new UserTouchHelper(0, ItemTouchHelper.LEFT, new UserTouchHelper.RecyclerItemTouchHelperListener() {
+        ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new FleetTouchHelper(0, ItemTouchHelper.LEFT, new FleetTouchHelper.RecyclerItemTouchHelperListener() {
 
             /**
              * callback when recycler view is swiped
@@ -62,7 +63,7 @@ public class UserFragment extends Fragment {
              */
             @Override
             public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction, int position) {
-                if (viewHolder instanceof UserAdapter.DataViewHolder) {
+                if (viewHolder instanceof FleetAdapter.DataViewHolder) {
                     // get the removed item name to display it in snack bar
                     String name = data.get(viewHolder.getAdapterPosition()).get("name");
 
@@ -75,7 +76,7 @@ public class UserFragment extends Fragment {
 
                     // showing snack bar with Undo option
                     Snackbar snackbar = Snackbar
-                            .make(UserFragment.this.getView(), " removed from list", Snackbar.LENGTH_LONG);
+                            .make(FleetFragment.this.getView(), name + " removed from list", Snackbar.LENGTH_LONG);
                     snackbar.setAction("UNDO", new View.OnClickListener() {
                         @Override
                         public void onClick(View view) {
@@ -91,7 +92,6 @@ public class UserFragment extends Fragment {
         });
         new ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(lst);
 
-
         load(loadJSONFromAsset());
 
         return v;
@@ -100,7 +100,7 @@ public class UserFragment extends Fragment {
     public String loadJSONFromAsset() {
         String json = null;
         try {
-            InputStream is = getActivity().getAssets().open("json/users.json");
+            InputStream is = getActivity().getAssets().open("json/fleets.json");
             int size = is.available();
             byte[] buffer = new byte[size];
             is.read(buffer);
@@ -114,6 +114,7 @@ public class UserFragment extends Fragment {
     }
 
     public void load(String jsonStr) {
+
         data = new ArrayList<>();
 
         try {
@@ -126,22 +127,14 @@ public class UserFragment extends Fragment {
                 HashMap<String, String> c = new HashMap<>();
 
                 c.put("type", "data"); // clasify the item if data or header
-                c.put("email", obj.getString("User.name"));
+                c.put("name", obj.getString("PluginFlyvemdmFleet.name"));
 
-                if(obj.getString("User.realname").trim().equalsIgnoreCase("")) {
-                    c.put("UserRealName", "without name");
-                } else {
-                    c.put("UserRealName", obj.getString("User.realname"));
-                }
-
-                if(!obj.getString("User.realname").equalsIgnoreCase("null")) {
-                    data.add(c);
-                }
+                data.add(c);
             }
 
             pb.setVisibility(View.GONE);
 
-            mAdapter = new UserAdapter(data);
+            mAdapter = new FleetAdapter(data);
             lst.setAdapter(mAdapter);
 
         } catch (Exception ex) {
