+++ app/src/main/java/org/flyve/admin/dashboard/ui/FragmentAbout.java
@@ -24,7 +24,7 @@
  * @link      https://flyve-mdm.com/
  * ------------------------------------------------------------------------------
  */
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -35,6 +35,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import org.flyve.admin.dashboard.R;
 import org.flyve.admin.dashboard.utils.EnvInfoAbout;
 
 public class FragmentAbout extends Fragment {
