+++ app/src/main/java/org/flyve/admin/dashboard/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.os.Bundle;
 import android.support.v4.app.FragmentManager;
@@ -11,6 +11,7 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import org.flyve.admin.dashboard.R;
 import org.flyve.admin.dashboard.adapter.DrawerAdapter;
 import org.flyve.admin.dashboard.utils.FlyveLog;
