+++ app/src/main/java/org/flyve/admin/dashboard/ui/SplashActivity.java
@@ -1,9 +1,11 @@
-package org.flyve.admin.dashboard;
+package org.flyve.admin.dashboard.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
+
+import org.flyve.admin.dashboard.R;
 
 public class SplashActivity extends AppCompatActivity {
