+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -2,6 +2,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.flyve.admin.dashboard.DashboardFragment">
+    tools:context="org.flyve.admin.dashboard.ui.DashboardFragment">
 
 </RelativeLayout>
